# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddressRequest:
    properties:
      city:
        example: Moscow
        type: string
      country:
        example: Russia
        type: string
      street:
        example: Vernadskogo
        type: string
    required:
    - city
    - country
    - street
    type: object
  dto.AddressResponse:
    properties:
      address_id:
        example: b1c4b160-b9e9-4f0c-83ec-9119e1f467f9
        type: string
      city:
        example: Moscow
        type: string
      country:
        example: Russia
        type: string
      street:
        example: Vernadskogo
        type: string
    type: object
  dto.ClientListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ClientResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 5
        type: integer
    type: object
  dto.ClientRequest:
    properties:
      address:
        $ref: '#/definitions/dto.AddressRequest'
      birthdate:
        example: "2006-11-25"
        type: string
      gender:
        enum:
        - male
        - female
        example: male
        type: string
      name:
        example: Egor
        type: string
      surname:
        example: Egorov
        type: string
    required:
    - address
    - birthdate
    - gender
    - name
    - surname
    type: object
  dto.ClientResponse:
    properties:
      address:
        $ref: '#/definitions/dto.AddressResponse'
      birthdate:
        example: "2006-11-25"
        type: string
      client_id:
        example: b1c4b160-b9e9-4f0c-83ec-9119e1f467f9
        type: string
      country:
        example: male
        type: string
      name:
        example: Egor
        type: string
      registration_date:
        example: "2026-06-01"
        type: string
      surname:
        example: Egorov
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
  dto.ImageRequest:
    properties:
      jpg_base64:
        example: iVBORw0KGgoAAAANSUhEUgAAB4AA...
        type: string
    required:
    - jpg_base64
    type: object
...
  dto.SuccessResponse:
    properties:
      status:
        example: created
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for shop service
  title: Shop API
  version: "1.0"
paths:
  /clients:
    get:
      description: Get clients list with optional params
      parameters:
      - description: Client name
        in: query
        name: name
        type: string
      - description: Client surname
        in: query
        name: surname
        type: string
      - default: 100
        description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClientListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get clients with params
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Create new client
      parameters:
      - description: Client data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create client
      tags:
      - clients
  /clients/{id}:
    delete:
      description: Delete client by id
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete client
      tags:
      - clients
  /clients/{id}/address:
    patch:
      consumes:
      - application/json
      description: Change address of client
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Address data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Change client address
      tags:
      - clients
  /images/{id}:
    delete:
      description: Delete product image by id
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete image
      tags:
      - images
    get:
      description: Get product image by image id
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get image
      tags:
      - images
    put:
      consumes:
      - application/json
      description: Change product image data
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      - description: Image data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Change image
      tags:
      - images
...
swagger: "2.0"